% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiwidgets.R
\name{widget}
\alias{widget}
\title{Create a generic HTML crosstalk user interface endpoint.}
\usage{
widget(class = c("transmitter", "receiver", "transceiver"), html,
  value = "value", ..., width = NULL, height = NULL)
}
\arguments{
\item{class}{The transmitter and receiever widgets use limited uni-directional crosstalk
communication. Transceiver widgets are more like normal crosstalk widgets.}

\item{html}{A valid HTML object string.}

\item{value}{The HTML object's I/O field name, for instance 'value' for some 'input' HTML elements.}

\item{...}{Additional arguments passed directly to the JavaScript code.}
}
\value{
An HTML widget object.
}
\description{
Use \code{widget} to define generic HTML elements with limited
crosstalk functionality. These elements can be used to control
crosstalk-enabled applications.
}
\note{
Typical differences with standard crosstalk widgets
\enumerate{
\item{Mostly use uni-directional communication, either transmitting or receiving data but not both.}
\item{Transmit raw values from the HTML object, not indexed shared data frame values.}
}
}

