% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiwidgets.R
\name{widget}
\alias{widget}
\title{Create a generic HTML crosstalk user interface endpoint.}
\usage{
widget(crosstalk, class = c("transmitter", "receiver", "transceiver"), html,
  value = "value", ..., width = NULL, height = NULL)
}
\arguments{
\item{crosstalk}{A crosstalk SharedData object. Currently only used for the crosstalk group.}

\item{class}{The transmitter and receiever widgets use limited uni-directional crosstalk
communication. Transceiver widgets are more like normal crosstalk widgets.}

\item{html}{A valid HTML object string.}

\item{value}{The HTML object's I/O field name, for instance 'value' for some 'input' HTML elements.}

\item{...}{Additional arguments passed directly to the JavaScript code.}
}
\value{
An HTML widget object.
}
\description{
Use \code{widget} to define generic HTML elements with limited
crosstalk functionality. These elements can be used to control
crosstalk-enabled applications.
}
\note{
Notable Features
\enumerate{
\item{Mostly uni-directional messaging to other widgets}
\item{Defaults to transmitting raw html object values (not indexed values from a shared data frame}
\item{Mostly uses the shared crosstalk data frame for its group key}
}
}
\section{Receiver-specific notes}{

Specify the optional \code{data=<vector>} argument to use indexed values in the
specified \code{data} vector indexed by the crosstalk key.
}

\section{Transmitter-specific notes}{

}

\section{Transceiver-specific notes}{

}

